requirements.txt
requests
faiss-cpu
sqlalchemy
tqdm
python-dotenv
redis>=4.0
python-docx